+++ src/Order/LineItem/Events/AfterIncItemMappingEvent.php
@@ -15,6 +15,8 @@ use Symfony\Contracts\EventDispatcher\Event;
  */
 class AfterIncItemMappingEvent extends Event
 {
+    /** Can be set in data bag to skip the item import */
+    public const SKIP = 'skip';
     private DataBag $mapping;
     private ShopgateOrderItem $item;
     private SalesChannelContext $context;